backend/requirements.txt
# FastAPI and Core Dependencies
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication & Security
pyjwt
passlib[bcrypt]
python-jose[cryptography]
cryptography

# HTTP & Web
httpx
requests
python-multipart

# Environment & Configuration
python-dotenv

# Background Tasks
celery
redis

# External Services
stripe
sendgrid

# Data Processing & Reports
pandas
numpy
reportlab

# File Handling
aiofiles
Pillow

# Template Engine
jinja2

# Utilities
python-slugify
python-dateutil

# Development & Deployment
gunicorn
